attestables/bob/pyproject.toml
[tool.poetry]
name = "bob"
version = "0.1.0"
description = ""
authors = ["Dhileane Rodrigues <dhileane.rodruiges@sou.unijui.edu.br>"]
readme = "README.md"
packages = [{include = "project"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest-cov = "^4.1.0"
flask-migrate = "^4.1.0"
dotenv = "^0.9.9"
flask-sqlalchemy = "^3.1.1"
sqlalchemy = "^2.0.40"
bcrypt = "^4.3.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
cryptography = "^44.0.2"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
isort = "^6.0.1"
autoflake = "^2.3.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "autoflake --in-place --exclude=*/migrations/* --remove-all-unused-imports -r . && black . && isort ."
pre_test = "task lint"
test = "pytest -s -x --cov-report=html --cov project -vv"